chore(LadderTest): LadderTest 파일 수정
